Added base components + moved functionality to other script.
