--- ginder/envconfig.ts
@@ -1,3 +1,0 @@
-import dotenv from 'dotenv';
-
-export const CLIENT_ID = process.env.CLIENT_ID || '';
