+++ ginder/src/lib/components/Footer/Footer.svelte
@@ -1,0 +1,9 @@
+<main>
+    <div id="footer">
+        <p>Created and maintained with 🤍 by <span><a href="https://github.com/boushrabettir">@boushrabettir</a></span></p>
+    </div>
+</main>
+
+<style>
+
+</style>
