+++ ginder/src/lib/components/Post/Post.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+
+</script>
+
+<main>
+
+</main>
+
+<style>
+
+</style>
