+++ ginder/src/routes/+page.svelte
@@ -1,14 +1,21 @@
 <script lang="ts">
-   
-    const CLIENT_ID = "";
+    import dotenv from 'dotenv';
+
+    const CLIENT_ID = process.env.CLIENT_ID || '';
     const login_with_github = () => {
         window.location.assign(`https://github.com/login/oauth/authorize?client_id=${CLIENT_ID}`)
     }
 </script>
 
 <main>
-    <p> Login through Github </p>
+    <h2 id="welcome">Welcome to...</h2>
+    <p> 👾 Login through Github 🤖 </p>
     <button on:click={login_with_github}>
         Login
     </button>
 </main>
+
+
+<style>
+
+</style>
